disseration-doc/loanpred/requirements.txt
absl-py==1.4.0
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
bleach==6.0.0
bokeh==3.1.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorcet==3.0.1
comm==0.1.2
contourpy==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
Django==4.1.6
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
editdistance==0.6.2
efficientnet==1.0.0
entrypoints==0.4
essential-generators==1.0
executing==1.2.0
flatbuffers==23.3.3
fonttools==4.38.0
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
holoviews==1.16.0
htmlmin==0.1.12
hvplot==0.8.3
idna==3.4
ImageHash==4.3.1
imageio==2.27.0
imgaug==0.4.0
inflection==0.5.1
ipykernel==6.19.4
ipython==8.7.0
itypes==1.2.0
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.1
jupyter_client==7.4.8
jupyter_core==5.1.1
keras==2.12.0
Keras-Applications==1.0.8
keras-ocr==0.8.9
keras-tuner==1.3.5
kiwisolver==1.4.4
kmodes==0.12.2
kt-legacy==1.0.5
lazy_loader==0.2
libclang==16.0.0
lightgbm==3.3.5
linkify-it-py==2.0.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
missingno==0.5.2
ml-dtypes==0.1.0
mlxtend==0.22.0
multimethod==1.9.1
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
pandas-profiling==3.2.0
panel==1.0.3
param==1.13.0
parso==0.8.3
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==4.22.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pyct==0.5.0
pydantic==1.10.7
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
pyviz-comms==2.3.0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.0
scipy
seaborn==0.12.2
shapely==2.0.1
simplejson==3.18.4
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tangled-up-in-unicode==0.2.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typing_extensions==4.5.0
uc-micro-py==1.0.2
uritemplate==4.1.1
urllib3==1.26.15
validators==0.20.0
visions==0.7.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.14.1
xgboost==1.7.5
xyzservices==2023.5.0
gunicorn
whitenoise